initialize with gaussian mixtures
I've got the problem cast in terms of an energy (cost) function.  But I
keep getting LinalgErrors.  The covariance matrix seems to be entering
forbidden regions presumably because it becomes non-invertible.
The standard reason for this is that the model is stuck in a component
with too few observations (If there is only one data point nearby the
eigenvalues will tend towards zero).